implemented algorithm
